requirements.txt
# Core API dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
starlette==0.35.1

# Authentication and security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2

# Rate limiting and middleware
slowapi==0.1.8

# ML model and dependencies
detoxify==0.5.1
transformers==4.22.1
torch==2.1.0
numpy==1.24.3
scipy==1.11.3
pandas==2.1.1

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Production utilities
prometheus-client==0.19.0
python-json-logger==2.0.7

# Health check dependencies
requests==2.31.0